+++ src/app/app.component.html
@@ -25,8 +25,8 @@
 		</mat-sidenav>
 
 		<mat-sidenav-content>
-			<div class="root-content" [@routeAnimations]="prepareRoute(outlet)">
-				<router-outlet #outlet="outlet"></router-outlet>
+			<div class="root-content" [@routeAnimations]="animation">
+				<router-outlet #outlet="outlet" (activate)="onActivate($event)"></router-outlet>
 			</div>
 		</mat-sidenav-content>
 		<mat-sidenav class="notif-sidenav" #snotif position="end" mode="over"
