+++ src/app/demo/pages/demo-export/demo-export.component.ts
@@ -5,9 +5,9 @@ import { BehaviorSubject } from 'rxjs';
 import { Item } from 'src/app/core/item/models/item';
 import { ItemService } from 'src/app/core/item/services/item.service';
 import { PageableDataSource, Paginator } from 'src/app/core/models/pageable-data-source';
+import { STAGE } from 'src/app/core/models/progress';
 import { ExcelExportService } from 'src/app/core/services/excel-export.service';
 import { SheetExportService } from 'src/app/core/services/sheet-export.service';
-import { STAGE } from 'src/app/models/progress';
 
 type LogItem = {
   timestamp: Date, message: string
