+++ src/app/shared/components/notification-center/notification.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { NotificationService } from 'src/app/core/services/notification.service';
 
-import { Notification } from 'src/app/models/notification';
+import { Notification } from 'src/app/core/models/notification';
 import { NotificationComponent } from './notification.component';
 
 describe('NotificationComponent', () => {
